READme.md
Image Recognition System
This project is a deep learning-based system that can accurately classify images into one of the 10 different categories in the CIFAR-10 dataset. It uses a convolutional neural network (CNN) architecture to extract important features from the images and then feeds these features into fully connected layers to make predictions. The system is trained on a subset of the CIFAR-10 dataset and achieves high accuracy on both the training and testing sets.

The project is implemented in Python using TensorFlow and can be easily adapted for other image recognition tasks. It takes input images of size 128x128 pixels, and the batch size can be adjusted for training.

Installation
To run this project, you need to install Python and TensorFlow. Clone this repository and navigate to the project directory in your terminal. Then run the command python main.py to start the image recognition system.

Usage
The system takes an input image of size 128x128 pixels and predicts its class label from 10 different categories in the CIFAR-10 dataset. The categories include airplane, automobile, bird, cat, deer, dog, frog, horse, ship, and truck.

License
This project is licensed under the MIT License - see the LICENSE file for details.

Acknowledgments
This project was inspired by the CIFAR-10 dataset and TensorFlow's image recognition tutorials. Special thanks to the TensorFlow team for providing the tools and resources to build this system.